move routes to model specific files
